requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
faker>=13.0.0,<14.0.0
watchdog>=2.1.0,<3.0.0
loguru>=0.6.0,<0.7.0
python-multipart>=0.0.5,<0.0.6
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=0.19.0,<0.20.0
aiofiles>=0.8.0